Convert card components to TypeScript (#1148)
These components were not outputting correct type definitions in the published
package because of an update to the TS version. Compilation worked correctly,
but consuming them did not. We also have to move away from the resolved aliases
for module imports because they do not compile correctly to their derived relative
paths.