Button onClick handler is optional for submit buttons (#1312)
I discovered this bug while working on converting a client component to TypeScript.
HTML buttons can be of type 'submit' inside of a form element. The form element can
have its own 'onSubmit' handler, which catches the click event of the child submit
button. In this case, a non-anchor/link button can still be a button element and not
require an 'onClick' handler. Update the component prop type to support this case.